フォームの一括更新機能

+++ app/controllers/payments_balances_controller.rb
@@ -1,6 +1,6 @@
 class PaymentsBalancesController < ApplicationController
   before_action :authenticate_user!
-  before_action :set_payments_balance, only: [:edit, :show, :update, :destroy]
+  before_action :set_payments_balance, only: [:edit, :show, :update]
   before_action :no_found_page, only: :show
 
   def index
@@ -23,49 +23,34 @@ class PaymentsBalancesController < ApplicationController
   def edit
   end
 
-  # def update
-  #   ids = params[:id].split(',').map(&:to_i)
-  #   attributes = params[:payments_balance][:payments_balances].values
-  #   if @payments_balance.update(ids, attributes)
-  #     redirect_to payments_balances_path(date: @selected_date)
-  #   else
-  #     render :edit
-  #   end
-  # end
   def update
-    # パラメータから削除するフォームのIDを取得
-    deleted_form_ids = params[:payments_balance].reject { |_, v| v.present? }.keys.map(&:to_i)
-  
-    # 削除したフォームのIDを用いて、該当するレコードをデータベースから削除
-    PaymentsBalance.where(id: deleted_form_ids).destroy_all
-  
-    # 残ったフォームの属性を更新
-    params[:payments_balance][:payments_balances].each do |index, attributes|
-      payment_id = attributes[:payment_id].to_i
-      date = attributes[:date]
-      purpose = attributes[:purpose]
-      amount = attributes[:amount].to_i
+    binding.pry
+    ids = params[:id].split(',').map(&:to_i)
+    attributes = params[:payments_balance][:payments_balances].values
 
-      # ここで該当するレコードを更新するか、新しいレコードを作成する処理を行う
-      # 例えば、以下のようにするとIDが存在する場合は該当するレコードを更新し、IDが存在しない場合は新しいレコードを作成します。
-      if index.present?
-        payments_balance = PaymentsBalance.find(index)
-        payments_balance.update(payment_id: payment_id, date: date, purpose: purpose, amount: amount)
-      else
-        PaymentsBalance.create(payment_id: payment_id, date: date, purpose: purpose, amount: amount)
+    # 既存フォーム削除
+    existing_form_ids = attributes.map { |nested_hash| nested_hash["id"].to_i }
+    deleted_form_ids = ids - existing_form_ids
+
+    # 新規フォーム作成
+    new_form_ids = Hash.new
+    unless new_form_ids.empty?
+      new_form_ids = Form::PaymentsBalanceCollection.new(payments_balance_params)
+      unless new_form_ids.save
+        render :edit
       end
     end
-  
-    redirect_to payments_balances_path(date: @selected_date)
-  end
-  
 
+    # 既存フォーム更新
+    if @payments_balance.update(existing_form_ids, attributes)
+      PaymentsBalance.where(id: deleted_form_ids).destroy_all
+      redirect_to payments_balances_path(date: @selected_date)
+    else
+      render :edit
+    end
+  end  
+  
   def show
-  end
-
-  def destroy
-    @payments_balance = PaymentsBalance.find(params[:id])
-    @payments_balance.destroy
   end
 
   private
@@ -84,6 +69,12 @@ class PaymentsBalancesController < ApplicationController
     @selected_date = Date.parse(params[:date])
     @payments_balance = PaymentsBalance.where(date: @selected_date, user_id: current_user.id)
     @payments_balances = PaymentsBalance.where(user_id: current_user.id)
+    # @balances = PaymentsBalance.find(params[:id])
+  end
+
+  def deleted_form(ids)
+    deleted_form_ids = PaymentsBalance.find(params[:id])
+    return deleted_form
   end
 
   def no_found_page
